Mainly interface changes in preparation to support incremental training
